+++ UI/_focus/Focus.ts
@@ -8,6 +8,7 @@
 import { detection, IoC } from 'Env/Env';
 
 import { collectScrollPositions } from './_ResetScrolling';
+import * as ElementFinder from "./ElementFinder";
 
 /**
  * make foreignObject instance. using for hack with svg focusing.
@@ -158,6 +159,30 @@ function makeResetScrollFunction(element: Element, enableScrollToElement: boolea
    return collectScrollPositions(element);
 }
 
+function matches(el: Element, selector: string): boolean {
+   return (
+      el.matches ||
+      el.matchesSelector ||
+      el.msMatchesSelector ||
+      el.mozMatchesSelector ||
+      el.webkitMatchesSelector ||
+      el.oMatchesSelector
+   ).call(el, selector);
+}
+function checkInput(el: Element): boolean {
+   return matches(el, 'input[type="text"], textarea, *[contentEditable=true]');
+}
+function getContainerWithControlNode(element: Element): Element {
+   while (element) {
+      // ищем ближайший элемент, который может быть сфокусирован и не является полем ввода
+      if (element.controlNodes && ElementFinder.getElementProps(element).tabStop && !checkInput(element)) {
+         break;
+      }
+      element = element.parentElement;
+   }
+   return element;
+}
+
 /**
  * Moves focus to a specific HTML or SVG element
  */
@@ -165,6 +190,29 @@ function focusInner(
       element: Element,
       cfg: { enableScreenKeyboard?: boolean, enableScrollToElement?: boolean } = {}
       ): boolean {
+   // на мобильных устройствах иногда не надо ставить фокус в поля ввода. потому что может показаться
+   // экранная клавиатура. на ipad в случае асинхронной фокусировки вообще фокусировка откладывается
+   // до следующего клика, и экранная клавиатура показывается не вовремя.
+
+   // можно было бы вообще ничего не фокусировать, но есть кейс когда это нужно:
+   // при открытии задачи поле исполнителя должно активироваться, чтобы показался саггест.
+   // но фокус на поле ввода внутри не должен попасть, чтобы не повторилась ошибка на ipad.
+
+   // поищем родительский элемент от найденного и сфокусируем его. так контрол, в котором лежит
+   // поле ввода, будет сфокусирован, но фокус встанет не в поле ввода, а в его контейнер.
+
+   // enableScreenKeyboard должен быть параметром метода activate, а не свойством контрола поля ввода,
+   // потому что решается базовая проблема, и решаться она должна в общем случае (для любого
+   // поля ввода), и не для любого вызова activate а только для тех вызовов, когда эта поведение
+   // необходимо. Например, при открытии панели не надо фокусировать поля ввода
+   // на мобильных устройствах.
+   if (!cfg.enableScreenKeyboard && detection.isMobilePlatform) {
+      // если попали на поле ввода, нужно взять его родительский элемент и фокусировать его
+      if (checkInput(element)) {
+         element = getContainerWithControlNode(element);
+      }
+   }
+
    const undoScrolling = makeResetScrollFunction(element, cfg.enableScrollToElement);
    const result = tryMoveFocus(element);
